Initial split of ImportComp from it's app.
