+++ install/bin/compimport.bat
@@ -1,1 +1,1 @@
-@$INSTALL_PATH\bin\decj decodes.tsdb.ImportComp %*%
+@$INSTALL_PATH\bin\decj decodes.tsdb.ImportCompApp %*%
